+++ app/src/main/java/app/nahehuo/com/bean/NetJobDetail.java
@@ -239,7 +239,7 @@ public class NetJobDetail {
             private String content;
             private String size;
             private String industry;
-            private int avgcomment;
+            private float avgcomment;
 
 
             public void setCid(int cid) { this.cid = cid;}
@@ -282,7 +282,7 @@ public class NetJobDetail {
             }
 
 
-            public void setAvgcomment(int avgcomment) {
+            public void setAvgcomment(float avgcomment) {
                 this.avgcomment = avgcomment;
             }
 
@@ -323,7 +323,7 @@ public class NetJobDetail {
             public String getIndustry() { return industry;}
 
 
-            public int getAvgcomment() { return avgcomment;}
+            public float getAvgcomment() { return avgcomment;}
         }
     }
 }
